Add Event Listener To Clear Completed Tasks
