+++ src/index.html
@@ -31,7 +31,7 @@
       </ul>
 
       <!-- Clear Completed -->
-      <a href="#" class="clear-completed">Clear all completed</a>
+      <span class="clear-completed">Clear all completed</span>
      </div>
     
    </section>
